Add all constants from NIST CODATA 2018, with an underlying redesign.
Each constant is now an object within the class Math::Constants::Constant. The attributes for each object are (where appropriate):

    name        -- the full name of the constant (as specified in the NIST file where appropriate)
    value       -- the value of the constant
    uncertainty -- the uncertainty in the value
    units       -- the units for the value
    symbol      -- the symbol for the constant where appropriate

Methods are defined which will emulate the previous interface where the constants were just reals -- these will return the
value of the constant as a string or Real as appropriate. The gist method will return a string containing the value, together with
the uncertainty and the units where appropriate.

In addition, export tags are now defined for all values. The constants which existed in previous versions of this module
have the DEFAULT tag. Thus a simple "use Math::Constants" will import exactly the same constants as before.

The math constants have a "math" tag, and the physics constants a "physics" tag. Thus only the appropriate constants can
be imported if desired.

In addition, the physics constants are split into sections depending on how they are defined:

    The fundamental physics constants (the seven constants which define the SI system) have the tag "fundamental".
        These constants are specified exactly and have an uncertainty of 0.
    The constants which have a defined value have the tag "defined". This includes the fundamental constants,
        as well as constants like the standard acceleration due to gravity.
    The constants which are derived from the defined constants have values here which use the formulae which
        define those constants rather than the less exact versions in the CODATA tables. They have a tag "derived".
    The constants which rely on measurements for their values have the tag "measured". Note that relationships
        between these values might not look exact. For example, the constant which is the ratio between the masses
        of the electron and proton might not be exactly equal to the ratio between the constants which represent those
        two values. This depends on the relative accuracy with which the ratio can be measured.